repurposed virtual_shopping_list.py
